requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
deep-translator==1.11.4
deepgram-sdk==2.12.0
distro==1.9.0
docopt==0.6.2
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
neuralspace==1.1.0
openai==1.6.1
pipreqs==0.4.13
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
tenacity==8.2.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.9.4
